requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-environ
psycopg2-binary
Pillow